+++ example/zod_validation_example.dart
@@ -8,14 +8,25 @@ void main() {
         'id': Zod().type<int>(),
         'name': Zod().min(3).max(10),
         'email': Zod().email(),
-        'phone': Zod().phone(),
+        'phone': Zod().optional(isValidWhenEmpty: false).phone(),
       },
     }
   };
 
   /// the received params from the request
-  final requestParams = <String, dynamic>{};
+  final requestParams = <String, dynamic>{
+    'user': {
+      'platform': 'web',
+      'data': {
+        'id': 1,
+        'name': 'John Doe',
+        'email': 'welito@gmail.com',
+        'phone': '',
+      }
+    }
+  };
 
   final result = Zod.validate(params: params, data: requestParams);
   if (result.isNotValid) print(result.result);
+  if (result.isValid) print('Valid');
 }