+++ lib/src/validations.dart
@@ -9,8 +9,9 @@ class Validations {
 
   /// verify phone validate
   static bool isPhone(String? phone) {
-    return RegExp(r"^\([1-9]{2}\) [2-9][0-9]{3,4}\-[0-9]{4}$")
-        .hasMatch(phone?.trim() ?? '');
+    final number = (phone ?? "").replaceAll(RegExp(r'[^0-9]'), '');
+    if (number.trim().length >= 10) return true;
+    return false;
   }
 
   /// verify required validate