+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+- Custom validate.
+- Optional validade.
+- Bug corrections.
+
+
 ## 1.0.0
 
 - Initial version.